Fixing typos in install.sh which broke mounting correctly
Their is a nasty typo in "skel/install/install.sh" much all versions going back to 1.9.1 (I don't fix older versions) 

When this is fixed, it exposes a second more obvious "typo" bug which was masked by the first.

This patch should be applied at to all the above branches, applies cleanly to head and trunk, but less cleanly to older branches where it should only be applied where it has an effect.

The second fix, is to skel/share/lib/namespace.sh this is very much dependent on how far Gerd has go in removing the need to mount PNFS. If the "dir" domain is needed to mount pnfs/chimera still then it shoudl be applied, which branches this should be applied to is dependent on Gerds knowledge.

Signed-off by: Paul
http://rb.dcache.org/r/808/