pool: Implement accept timeout for xrootd
The old xrootd implementation had a hardcoded timeout for how long to
wait for a redirected client to connect to the mover. The initial
version of the new xrootd mover did not have this timeout.

Without this timeout, it is quite common that pools get overloaded with
xrootd movers that are never activated.

Target: trunk
Target: 1.9.5
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/827/