SRM: Remove deadlock caused by implicitly calling job's toString() method in lo$
Remove the deadlock caused by implicitly calling job's toString() method
in log statements of the SharedMemoryCache. Job.getJob (syncronization
on the weakHashTable) can call methods of the SharedMemoryCache
(synchronization on the Reentrant Lock) , which inside called toString,
which for a file request may call methods that try to obtain the objects
assosiated with requests which calls getJob for the request. If done in
two threads, this can lead to a deadlock. By removing the calls in the
SharedMemoryCache that can lead to a deadlock as described in rt5153.
This patch removes this possibility.

Target: trunk
Target: 1.9.5
Signed-off-by: Timur Perelmutov <timur@fnal.gov>  
Acked-by: Gerd Behrmann <behrmann@ndgf.org>
Patch: http://rb.dcache.org/r/825/
Ticket: #5153