pool: Fix recovery of BROKEN files with Berkeley DB backend
The error state is persistent when the Berkeley DB backend is used. This
caused the error recovery in ConsistentStore to be skipped. This patch
changes the state from BROKEN to FROM_CLIENT to force a replay of the
registration which happens after a file has been received.

In addition the patch changes the call to PnfsManager from the old
addCacheLocation and setFileSize operations to setFileAttributes.

Target: trunk
Target: 1.9.5
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/826/