PnfsManager: Make PnfsGetFileMetaDataMessage a subclass of PnfsGetFileAttributes
The ultimate goal is to get rid of FileMetaData and StorageInfo and thus
also PnfsGetFileMetaDataMessage and PnfsGetStoraInfoMessage. One
complication is that PnfsCreateEntryMessage inherits from
PnfsGetStorageInfoMessage.

To transition to FileAttributes, this patch makes
PnfsGetFileMetaDataMessage inherit from PnfsGetFileAttributes. Thus it
is up to the sender to decide between using FileAttributes or
FileMetaData and StorageInfo when creating new entries. Once all
components have been transitioned to FileAttributes, we remove
PnfsGetFileMetaDataMessage and PnfsGetStorageInfoMessage and let
PnfsCreateEntryMessage inherit from PnfsGetFileAttributes directly.

One thing which makes this transition even more complicated is that
pools send PnfsGetStorageInfoMessage to PnfsManager. Thus some backwards
compatibility is required. This patch solves this issue by

a) Maintain all existing fields of PnfsGetFileMetaDataMessage and
   PnfsGetStorageInfoMessage. Thus a 1.9.5 pool can understand the reply
   of a 1.9.6 PnfsManager.

b) Override getRequestedAttributes in PnfsGetFileMetaDataMessage and
   PnfsGetStorageInfoMessage such that missing information is filled in.
   This allows a 1.9.6 PnfsManager to understand the request from 1.9.5
   pool.

In 1.9.7 we can update this code to remove fields in
PnfsGetFileMetaDataMessage and PnfsGetStorageInfoMessage. Thus we
maintain compatibility between 1.9.5 and 1.9.6 and between 1.9.6 and
1.9.7. Compatibility is however not transitive.

PnfsManager no longer needs to process PnfsGetFileMetaDataMessage and
PnfsGetStorageInfoMessage explicitly, as they can be handled by the same
code processing PnfsGetFileAttributes.

The FileMetaData class got a new constructor for converting a
FileAttributes instance.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/829/