Quick and dirty improvement of exception.
PNFS manager has crashed due to power outage while Cleaner cell was active.

Upon restart Cleaner is not starting, and there are the following errors in PnfsManager log:

06/10 11:49:00 Cell(cleaner@pnfsDomain) : Problem in getting global lock : operation locked

The file GlobalLock is present in "current" directory and has a timestamp of the time of crash.

This bug fix changes the exception to 

06/10 11:49:00 Cell(cleaner@pnfsDomain) : Problem in getting global lock : File Exists : /absolute/path

Signed-off by: Gerd
http://rb.dcache.org/r/768/