Adding an rpm dependency to SRM server and SRM client RPM's on java > 1.5 so the client can be deployed on glite 3.2.
Because the glite UI/WN on SL5 does not include a dependency on j2sdk either java-1.6.0-sun-compat or java-1.5.0-sun-compat are not installed, so making gcj the default java for the UI.

Fortunately gcj does not pretend to be a java 1.5 implementation which ties in nicely with our dependencies. Also this choice of dependacy is now set to conform with fedoraproject.org standards.

http://fedoraproject.org/wiki/Packaging/Java#BuildRequires_and_Requires

Signed-off by: Tigran
http://rb.dcache.org/r/839/