pool: Fix sticky bit expiration
It used to be the case that sticky bits would be removed as a
side-effect of querying the list of sticky bits. This was however
changed some time ago, as the side effects was not desirable (it had
side effects!).

This however means that the expiration timer logic is wrong. Previously
we would only schedule a timer for the expiration bit that would live
the longest. The point being that other expiration bits that died before
would be removed when queried and would thus not be visible to the user.

This patch updates the logic such that we schedule an expiration timer
for the sticky bit that will expire first. When the timer is triggered,
we schedule a new expiration task.

I also added a 1 second delay from the expiration point till the timer
is triggered. I did this because we observed that sticky bits would
sometimes not expire on our pools. The only explanation I could find was
if the timer would trigger just before the sticky bit actually expired.

The problem is present in 1.9.3 and 1.9.4 pools, too. The sticky bits
are however purged by a pool restart so I do not consider the bug
urgent enough to commit to those branches.

Target: 1.9.5
Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/857/