cells: Unify interpretation of ${0} when doing exec
Before this patch, ${0} would be the file/variable name specified on the
exec command line, except when the -loop option was used. If the -loop
option was used, then ${0} was the first argument.

This patch eliminates this inconsistency by shifting the arguments when
giving to exec. Thus ${0} is now always the first argument to the block
being executed.

COMPATIBILITY WARNING: Custom batch files that happen to use parameters 
to non-loop exec commands must be updated.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/861/