Catch any Throwable thrown by a scheduled task in the ScheduledExecutorService in StateMaintainer
This patch updates the StateMaintainer so that any Throwable that was thrown by a scheduled tasks is 
suitably logged.

The ScheduledExecutorService (part of the java.lang.concurrency package) wraps all tasks within a 
ScheduledFuture object.  This captures any Exception, so preventing an UnhandledExceptionHandler 
from seeing them.  Calling the get method of the Future will expose any thrown Exception (by 
throwing an ExecutionException).

However, if we want to "fire-and-forget" the tasks, we do not keep the Future objects.  Altering 
behaviour to keep track of the Future simply to discover if an exception has been thrown is both 
awkward and fragile.  The solution proposed here is to wrap tasks within a new Class (called 
FireAndForgetTask) that catches any Exception (indeed, any Throwable) and calls the Thread's 
UnhandledExceptionHandler.  By default, within dCache, this will propagate to the JVM-wide 
defaultUnhandledExceptionHandler.  This is set in the SystemCell to log these exceptions using 
log4j.

The FireAndForgetTask is placed in opt.dcache.utils package as a very similar class is an inner 
class used within the pool migration module. (The similarity is not an accident.)  The hope is that 
the pool-specific implementation will be replaced by a usage of FireAndForgetTask with a future 
patch.

The patch does the following ancillary work:
  1. clarifies exactly when the StateMaintainer monitor is entered.
  2. It fixes an issue where a purge-metrics task may be scheduled when an existing task is underway.
  3. It passes the correct ThreadFactory to the singleThreadScheduledExecutorService.

Acked-by: Gerd
RB: http://rb.dcache.org/r/872/