PoolManager: Always enable watchdog
The watchdog is essential to guarantee that PoolStatusChanged messages
are sent when pools go down. However we did not enable it by default. In
fact, dCache.batch didn't even show the options to enable it and I doubt
any documentation mentioned that it existed.

This patch changes the initialization logic such that the watchdog is
always enabled; if not configured explicitly then default values are
used.

Target: trunk
Target: 1.9.5
Target: 1.9.4
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/876/