pool: Handle UnresolvedAddressException in FTP mover
The FTP mover may throw UnresoledAddressException when it fails to
resolve the name of the host it should connect to. This exception would
trigger a rather useless message to be sent to the client:

426 426 Transfer aborted (Transfer failed due to unexpected exception: null)


This patch changes this to a more meaningful message.

Target: trunk
Target: 1.9.5
Acked-by: Timur Perelmutov <timur@fnal.gov>
Patch: http://rb.dcache.org/r/846/