Downgrade an info SIP logging message since it isn't an error
The message "Inconsistency in information: pool <poolname> accessable via link <linkname> but not 
present under poolspace" indicates that a pool is mentioned in a link but is (apparently) not 
known to the pool manager as a pool.  This would indicate some inconsistency.  However, this can 
happen temporarily due to how info service collects the information; e.g, when info is 
boot-strapping its initial set of information or when the set of pools that poolmanager knows 
about changes.

This patch changes this message from a warning to debugging information.  The wording is also 
improved slightly.

Acked-by: Gerd
RB: http://rb.dcache.org/r/879/