xrootd: Initialize request ID in pool request message
Pools can collapse or discard duplicate messages based on the request ID
of the transfer request message. Previously the xrootd door would not
initialize this field. As a consequence collapsing of identical requests
would not work, and the pool would generate a ton of warnings about
duplicate requests.

This patch changes the door such that the xrootd file handle is used as
the request id. This handle is unique for an open request throughout the
lifetime of the door. Thus different open calls will using different
request id, however internal timeouts and resubmissions will cause the
same request id to be used.

Target: trunk
Target: 1.9.5
Target: 1.9.4
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/887/