Fix info service SIP NAS naming and PaintInfo equals
This patch fixes two related problems.  The first problem is that the
equals method of the PaintInfo (incorrectly) depends on the set of
links that the corresponding pool is accessible through, but the name
of the NAS, used to group pools with the same accessibility,
(correctly) doesn't depend on the links.  Some error-checking code
caught this inconsistency, resulting in a RuntimeException being
thrown.  The fix is to remove the dependency on the set of links from
the hashCode and equals methods.

The other problem was that the NAS names were generated in a very
primitive fashion: taking the hashCode of the _access object.  This
has many issues, not least of which is that it introduces the
possibility (albeit a vague one) of NAS name collisions.  This patch
addresses that by calculating a descriptive but terse name for the
NAS.  The unit tests have been updated, so give examples of these
terse names.

With the NAS name calculated more appropriately, the field _access
became redundant.  Therefore, this patch completely removes it.

Acked-by: Gerd
Acked-by: Tigran
RB: http://rb.dcache.org/r/884/