pool: Fix file system size check in case Java doesn't know the size
On Solaris the 32-bit Java (which is the default) cannot report the
size of large file system. Instead it silently reports 0. This
obviously breaks our file system size check which we added in 1.9.5.

This patch adds a work around that causes the check to be skipped in
case the file system size is reported as 0.

Target: trunk
Target: 1.9.5
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/890/