add Obsoletes: dcache-dcap in new dcap packaging spec files.
This change makes the new form of packaging dcap replace the old form of packaging dcap smoothly on upgrade for non glite sites.

This will make life easier for sites like NAF etc. This will not be an issue for glite sites as this is done at the metapackage level.

Signed-off by: Tigran
http://rb.dcache.org/r/892/