cells: Protect cell against SerializationException
If a cell would encounter a SerializationException when deserializing a
message, it would die. This patch changes the logic such that the cells
logs a message at log level FATAL, but otherwise continues to run. This
avoids that a single broken cell in the system can bring down the
complete dCache.

Handling of other uncaught exceptions has been changed such that the
uncaught exception handler of the thread is invoked. This ensures that
OOM errors trigger a restart.

Target: trunk
Target: 1.9.5
Target: 1.9.4
Ticket: #5178
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/918/