pinmanager: Adjust timeout settings
I was again woken in the middle of the night by a Nagios alarm. Our test
file was once again "stuck" inside the PinManager. The problem was
triggered by a temporary connection outtage to one of the pools
containing the file (we got several copies of the test file). PinManager
logs showed that the PoolManager had returned a read pool to the
PinManager, but just as the PinManager sent a SetStickyMessage to the
pool, the connection was dropped for a minute or two. As a result the
reply from the pool was never received by the PinManager.

This would not have been so bad if it had simply caused a single failed
test. However, the PinManager collapses pin requests to the same file.
As it also got very conservative timeout settings, this caused all
subsequent attempts to access the file to fail too (the client timed
out). I tried to force an unpin, but that didn't work either (don't know
why - couldn't reproduce it in my test setup).

This patch adjusts the timeout settings in the PinManager. The new
values are IMO still quite conservative. The settings are:

- PoolManager is unchanged at 1 hour. For disk only files this is way to
much. For tape files this is not enough. This is an architectural
problem that we have to solve later.

- PnfsManager is reduced from 1 hour to 5 minutes. IMO if PnfsManager
takes 5 minutes to respond, then the system is in big big trouble. 5
minutes is the timeout we have in several doors. In particular with
Chimera I feel it is still too conservative, but I can live with 5
minutes for now.

- Pool is reduced from 1 hour to 90 seconds. Pools should really respond
quickly to all requests. If they don't, then another pool should be
selected.


Target: 1.9.5
Target: trunk
Acked-by: Timur Perelmutov <timur@fnal.gov>
Patch: http://rb.dcache.org/r/896/