xylophone: fix bugs in attr support for dynamic, repeated attributes
LDAP supports repeated attributes.  Xylophone includes support for this by allowing one to
declare an attribute repeats in one of two ways: either based on a fixed list of items or
based on the supplied dynamic XML (which, for us, is the XML file produced by the info
service).

The former option (repeated LDAP attributes based on a list) is used in the current info
provider (glue-1.3.xml.template) whilst the latter (repetition dictated by the dynamic XML
file) isn't; however, repeated attributes from the dynamic XML file will be needed by the
forthcoming SpaceToken+InstalledCapacity info provider.

It turns out that there is a bug in xylophone that, when repeating using the XML file with
a relative XPath, the wrong number of attributes are emitted and they have the wrong
values.  (So, there's two bugs being fixed)

Although the bug in xylophone is present in all branches, this patch is only for 1.9.5 and
trunk since only those branches need the (currently broken) functionality.

Acked-by: Gerd
RB: http://rb.dcache.org/r/899/