xylophone: fix bug preventing use of list-iterator item from evaluation within a predicate
The two predicate elements (allow and suppress) allow for dynamic control over which
objects are published in xylophone.

A bug, present in all current versions of dCache, prevents the use of the list iterator
item when evaluating whether an object should be published.  This patch fixes this bug.

In keeping with the last patch, this patch is only for 1.9.5 and trunk.

Acked-by: Tigran
RB: http://rb.dcache.org/r/924/