Update info provider to publish space tokens and installed capacity information
This patch represents a major overhaul to the info provider.

It provides GlueSA objects for each NAS.  A NAS is a non-overlapping group of pools with
consistent VO accessibility.  It publishes the appropriate VOInfo child objects for these NAS
GlueSAs to allow non-SRM clients to discover path information.  These GlueSA objects also allow
"installed capacity" accounting of a dCache instance: where capacity and usage is reported, both
in total and per VO.

GlueSA objects are also generated for each SRM space reservation description grouping.  The
aggregate information for SRM reservations with the same description (for the same VO) are
published; although, for most sites, each such grouping will likely have only one SRM
reservation.

The patch replaces the existing glue-1.3.xml.template file as we want new sites to use this info
provider by default.  However, not described in the patch (but should be considered part of it),
the existing glue-1.3.xml.template file will be renamed glue-1.3-old.xml.template.  This is to
allow sites to continue using the existing info provider should they find problems with the new
one.

Please note there a couple of known issues with the current system; specifically, 1) suppressing
a unit-group's mention within a GlueSA will result in ACBR attributes that don't have a named VO
(like "VO:" instead of "VO:atlas"), 2) SRM reservation-groups (those reservations with the same
description and for the same VO) are currently assumed to have the same access-latency and
retention-policy.

Due to how SRM and GLUE is used, neither of these issues are believed cause a problem, but they
will be fixed in follow-up patches.

Acked-by: Gerd
RB: http://rb.dcache.org/r/929/