srm: Use new PermissionHandler in SRM
Some of the SRM code performs custom permission checking, some delegates
to the PnfsManager and some uses the classic pre-ACL PermissionHandler.

This patch replaces some of the uses of the classic PermissionHandler
and custom permission checking with the use of the new FileAttributes
based PermissionHandler. Ideally we would always delegate permission
checking to PnfsManager, however this is not always possible within the
existing infrastructure of the SRM. The remaining cases should use the
new PermissionHandler and fetch FileAttributes to check permissions
(including ACLs).

There is still plenty of code left doing it the "old way", however to
keep the patch size down, not everything is updated in one patch.

This patch was extract from patch 840.

Target: trunk
Patch: http://rb.dcache.org/r/845/
Acked-by: Vijay Sekhri <sekhri@fnal.gov>