nfs: mover: keep track of space allocation
Each NFS request has a session is which is associated with
a 'mover'. The 'classic' mover is only to make it work inside a pool.
While each nfs operation is state less there is no information
how much disk space was allocated for particular file. The Allocator
do not provide such information, and as a result of it each mover does it
internally. This patch adds a similar book keeping into nfs mover.


Target: trunk, 1.9.5
Acked-By: Gerd
Patch: http://rb.dcache.org/r/922/