srm: Transition SRM from PnfsGetFileMetaData/StorageInfo to PnfsGetFileAttributes
This patch updated the DcacheFileMetaData class to use FileAttributes
rather than FileMetaData+StorageInfo. The patch moves the conversion of
the dCache classes to DcacheFileMetaData from the Storage class to a new
constructor of DcacheFileMetaData.

The patch updates the Storage class and the various companions
accordingly.

Storage.setFileMetaData is updated to use PnfsSetFileAttributes. Notice
that PnfsManager does not allow updates on access time and modification
time (such updates would always be ignored) and thus the patch removes
this part of the method.

Notice that the new code in several cases delegates permission checking
to the PnfsManager. This has one side effect that may or may not be
desirable: Fetching StorageInfo is considered as a "read" operation on
the file and permissions are checked accordingly. This works fine
regular doors, but the SRM uses Storage.getFileMetaData as a universal
method to get information about a file, no matter whether it intends to
read it or not. If the user doesn't have read permission on the file,
this lookup will fail.

There are two places that need to be fixed to resolve this issue:

First, PnfsManager cannot really distinguish between the equivalents of
a stat() and an open() in read-mode. We may need to add
PnfsReadEntryMessage as a subclass of PnfsGetFileAttributes with the
only difference that PnfsReadEntryMessage will fail if read permissions
are not allowed. Then we can change PnfsGetFileAttributes such that it
will only check whether the subject is allowed to stat the file.
Alternatively we can update PnfsGetFileAttributes such that it
dynamically restricts which information is included in the reply if the
subject doesn't have permission to read all attributes. This however
still has the problem that the SRM would not be able to fetch
StorageInfo unless the user had read permissions on the file.

Second, the SRM itself suffers from the same problem: The Storage class
cannot distinguish whether getFileMetaData is a precursor to reading the
file, or simply a 'stat' operation. The Storage class does provide
getGetTurl methods that could be used for permission checking, however
this is not how it is used today. Thus this requires considerable
restructuring of the interface and the SRM code.

One of the benefits of the patch is a faster Storage.getFileMetaData
implementation as we cut the extra roundtrip to PnfsManager to fetch the
file locations.

The patch is an essential step towards speeding up SRM ls: The key to a
fast list operation is to fetch FileAttributes inline with the list
operation. This is support by PnfsManager, but until now this didn't
help with the SRM as it didn't understand the FileAttributes class.

Target: trunk
Acked-by: Timur Perelmutov <timur@fnal.gov>
Patch: http://rb.dcache.org/r/942/