PoolManager: Stop ping handler on error
When the PingHandler times out, we forgot to stop it. That caused it to
keep timing out with an error even though we had already given up on the
stage or p2p.

I'm not sure whether this caused any problems besides the misleading
error messages.

This patch adds a call to stop() in case of errors.

Target: trunk
Target: 1.9.5
Target: 1.9.4
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/960/