PnfsManager: Fix permission checking for list and lookup
For some reason PermissionHandlerNameSpaceProvider did not contain the
necessary code for permission checks for list operations - even though
the JavaDoc said that it should. This patch adds this code.

I also noticed that several of the existing checks resolved the path to
PNFS ID. When they do that, they called pathToPnfsId of the parent
class, thus bypassing the lookup permission check. However if the
operation is done by path, then we have to verify that the user has
permissions to resolve the path. Therefore this patch updates most of
the pathToPnfsId calls to use the version that performs permission
checking.

Target: trunk
Target: 1.9.5
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/962/