PnfsManager: Add workaround for negative directory entry cache problem
The following fails with PNFS on some versions of Linux:

for ((i=1; i<=10; i++)); do
   cat ".(id)(newdir$i)"
   mkdir newdir$i
   cat ".(id)(newdir$i)"
done

The problem is believed to be caused by a negative cache entry created
by the first cat.

The above sequence happens inside dCache when a new directory is
created. When the problem occurs, the result is a root owned directory.

This patch adds two workarounds to the PnfsFile.getPnfsId method. First, 
we check that the name space entry exists before attempting to retrieve 
the PNFS ID. Second, if the lookup fails, we attempt to create the 
.(id)(...) file. This operation will fail, but it has the side effect of 
flushing the name from the negative directory entry cache. Afterwards we 
try to read the PNFS ID a second time. Only if that fails too will we 
conclude that the file does not exist.

Target: trunk
Target: 1.9.5
Signed-off-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Acked-by: Gerd Behrmann <behrmann@ndgf.org>
Patch: http://rb.dcache.org/r/979/
Patch: http://rb.dcache.org/r/976/
Ticket: #5212