http://rb.dcache.org/r/930/, http://rb.dcache.org/r/932/, http://rb.dcache.org/r/933/, http://rb.dcache.org/r/934/, http://rb.dcache.org/r/935/, http://rb.dcache.org/r/936/, http://rb.dcache.org/r/937/, http://rb.dcache.org/r/938/, http://rb.dcache.org/r/940/
srm: refactor srm request execution code synchronization to make it compatible with terracotta 

Terracotta requires that access to all fields that belong to clustered objects are synchronized.  The recommended way of synchronization is using Reentrant Read Write Locks. This series of patches introduces two methods rlock and wlock that perform locking using the Reentrant Read Write Locks and two corresponding methods runlock and wunlock for unlocking. All code that needs to be synchronized is preceded by lock statement,  included in a try block, and is followed by the finally with corresponding unlock method inside.  I will refer to this as synchronized block in my further explanation, even though it is not traditional synchronized statement.

In many places the direct access to the variables is replaced with bean pattern methods. Variables that do not change are declared final and therefore access to them needs no synchronization. 

In addition to mechanically changing the code to using the new synchronization, the patches also eliminate locking where it is not needed and change code to eliminate sources of potential deadlocks. Since the file requests require access to the request fields the locking is always performed on the file requests first and then on requests. The requests do not perform file requests operations in the synchronized blocks.