info-provider: add support for generating LDIF for a site-level BDII
Most sites have a resource-level BDII for their storage element.  This accepts
information from the info-provider, which the site's single site-level BDII will
update itself from periodically.

It is technically possible for the site-level BDII to accept data from the
info-provider directly.  However, the site-level BDII requires LDIF with a slightly
different structure (the parent object in the DIT is different).  Because of this,
the output from the info-provider cannot be used directly.

This patch adds support for the info-provider producing LDIF suitable for injecting
into a site-level BDII.

Acked-by: Gerd
RB: http://rb.dcache.org/r/985/