info-provider: work-around bug in how FTS obtains list of supported VOs
The SRM interface MUST be published in GLUE in two places: as a
GlueSEControlProtocol and as a GlueService object.  One of the reasons for
publishing the object twice is that the GlueService object allows a site to publish
which VOs are allowed use a particular SRM end-point.

FTS makes use the fact that GlueService objects can specify the allowed VOs.  It
builds a list of supported VOs for each end-point.  When considering a transfer to
some site, it will only broker transfers to some specific SRM end-point if it knows
that the VO is supported.

Unfortunately, FTS builds the list of supported VOs from those published as
GlueServiceAccessControlRule or GlueOwner attributes.  GlueServiceAccessControlRule
is completely undocumented attribute (in GLUE) and is marked deprecated in the LDAP
schema.  It seems that this attribute has been deprecated since 2004 (if not
earlier). The GlueOwner attribute is poorly documented in Glue and is currently
described in EGEE documentation as "not used".

Moreover, FTS completely ignores the well-documented method of expressing allowed
VOs: the GlueServiceAccessControlBaseRule attribute.  This has been reported in
Savannah:

  https://savannah.cern.ch/bugs/?57827

This patch adds a work-around to allow FTS to continue functioning.  The work-around
is to publish the GlueServiceAccessControlRule attributes in addition to publishing
the correct GlueServiceAccessControlBaseRule.

Acked-by: Gerd
RB: http://rb.dcache.org/r/987/