info-provider: fix publishing of SRM so (single) aliased FQDNs are supported
A number of sites use an alias for their SRM; for example, a machine 
dcache01.example.org might actually provide the SRM end-point but, to avoid tying 
the service to that particular machine, the site may wish to publish their SRM 
end-point using the DNS name srm.example.org, which maps to the same IP address as 
dcache01.example.org.

The current info-provider does not allow for this: it publishes all doors with the 
FQDN of their (public) interface.  For the above example, this would be 
dcache01.example.org.

Some component of dCache needs to know that dcache01.example.org should be presented 
as srm.dcache.org.  Currently, the only component that can do this is the 
info-provider.  Therefore, support for this is introduced here.  This may be part of 
a bigger issue of how a door presents itself to clients.

The patch makes use of a current naming convention: the GlueSEUniqueID is required 
to be the FQDN of the SRM end-point.  Rather than requiring the site-admin to 
configure the SRM end-point's FQDN a second time, the info provider simply uses the 
GlueSEUniqueID to build the URI of the SRM end-point.

Acked-by: Gerd
RB: http://rb.dcache.org/r/986/