info-provider: improve the work-around for BDII bug.
A previous patch-set (12704, 12705) introduced a work-around for a bug in BDII.  Unfortunately, 
this work-around was incomplete as it mapped only one of the characters that RFC 2253 requires to 
be escaped to a safe character.  NAS names can also include semi-colon characters that, too, 
require escaping under RFC 2253.

This patch fixes this issue by mapping all such characters to the colon character.

Acked-by: Gerd, Tigran
RB: http://rb.dcache.org/r/996/