info-provider: suppress VOInfo objects without a VO
In Glue 1.3, the GlueSA object takes zero or more VOInfo objects.  The VOInfo object provides a
VO-specific view of the space represented by a GlueSA and represents the VO's ability to write
into that space; for example, the path that a VO should use is described in the VOInfo object.

The site admin may choose to suppress publishing information on certain store units.  They may do
this by specifying an empty path (to prevent publication of the VOInfo object) or an empty VO (to
prevent any mention in the GlueSA object).

The current info provider has a bug where, if the site admin suppresses the store unit by
specifying the corresponding VO is empty but that the path takes a default value (/UNDEFINEDPATH)
then the VOInfo object is published, albeit one without any VO.

This is pointless and confusing: there is no point in publishing a VOInfo object if there is no
corresponding VO: no VO can select for this VOInfo object, making it redundant.

This patch fixes the info provider by suppressing the publication of a VOInfo object if the store
unit has no corresponding VO.

Acked-by: Tigran
RB: http://rb.dcache.org/r/992/