xylophone: fix lookup so they use the default value when XPath result is empty
The lookup element in Xylophone supports the concept of a default value.  This default is either 
a simple value (when expressed as an attribute) or an expansion (when path is an attribute).  If 
the lookup "fails" then the default value should be used.  Fails, here, means the XPath selects 
an empty node-set: the XPath is still required to be valid.

Recent changes that introduced support for XPath-based lookup element's path broke support for 
default values.  With the current version, a failed expansion results in an empty string being 
used.

This patch restores correct behaviour.

Acked-by: Gerd
RB: http://rb.dcache.org/r/994/