core: Let CellStub propagate NoRouteToCellException
CellStub.send would silently drop NoRouteToCellException. The idea being
that using CellStub.send indicates a one-directional communication and
in that case the sender has to be prepared that messages are silently
lost (as that cannot be prevented).

However the name CellStub.send is misleading in this respect as it
doesn't really conway that this is about one way communication. In
addition to this issue, CellStub cannot provide meaning full logging of
a NoRouteToCellException in this case, as it doesn't know the severity
of loosing this message.

This patch changes the interface of CellStub.send such that
NoRouteToCellException is propagated. Callers are updated to log the
event.

PnfsHandler.send is changed in the same way as it suffers from the same
usability flaw. PnfsHandler.notify is added for the send-and-forget
case.

Target: 1.9.5
Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1022/